+++ WebApiTerra800/Properties/AssemblyInfo.cs
@@ -30,5 +30,5 @@ using System.Runtime.InteropServices;
 //
 // You can specify all the values or you can default the Revision and Build Numbers 
 // by using the '*' as shown below:
-[assembly: AssemblyVersion("0.6.640.0")]
-[assembly: AssemblyFileVersion("0.6.640.0")]
+[assembly: AssemblyVersion("0.6.650.0")]
+[assembly: AssemblyFileVersion("0.6.650.0")]